scripts: extract shared wasm_exec.js transform logic
Move the common AST patching steps out of transform-go.ts and
transform-tinygo.ts into a shared transformWasmExec() helper, parametrized
by globalName, expected standalone-global count, and the globalProxy decl.
TinyGo keeps its Node.js-block commenting pre-step locally and delegates
the rest. Output for Go 1.26.2 / TinyGo 0.41.1 is byte-identical to before.
